Fix raw selectors.
Raw selectors weren't working because the logic to determine if the selector was using xpath or css was off. The `evaluate` method can use the valid css selector however it returns an empty `NodeList`. There are times when the `evaluate` method can't parse the given selector and we just recieve an empty array. `TypeBase` was treating `h1` as a valid xpath and as it's not valid xpath it would return an empty NodeList.

Fixes #1