Introduces Fetcher classes, local cache and on-load JS
Fetcher classes facilitate fetching the URL content.  The built-in Fetchers feature the old RollingCurl methods, a new and updated multi curl class and a class that utilises the Spatie Crawler.  This change also brings with it a local disk cache which is transportable and can be used to re-run a migration at a later date.  It also facilitates a basic resume as cached content is used preferentially and any non-cached urls in the configuration will be fetched in the run.