Feature/spatie fetch content, cache, JS, fetch and url tests (#77)
* Spatie crawl instead of rollingcurl

* WIP but functional implementation of local caching of url content, replaces rolling curl with Spatie crawler and adds JS support.  Work towards #41 and #71.

* Introduces Fetcher classes, local cache and on-load JS

Fetcher classes facilitate fetching the URL content.  The built-in Fetchers feature the old RollingCurl methods, a new and updated multi curl class and a class that utilises the Spatie Crawler.  This change also brings with it a local disk cache which is transportable and can be used to re-run a migration at a later date.  It also facilitates a basic resume as cached content is used preferentially and any non-cached urls in the configuration will be fetched in the run.

* Move default settings into container class

* Expose ingore_ssl_errors option in fetch_options

* Clearer text label

* phpcs

* Ignore SSL change

* Remove duplicate key

* Allow setting user agent as an option and update docs

* Fix ups and start of tests

* Fix ups and start of tests

* Remove oncomplete and implements tests

Tests the major fetching options specified via fetch_options and url_options, including duplicate content, query, fragment, javascript, user-agent, redirects

* Fix doc var name

* Added puppeteer and chromium.

* Remove dependencies.

* Restrict phpcs to src folder.

* Fix localhost assertion.

* Added --no-cache flag.